+++ src/scheduled-query/azext_scheduled_query/_params.py
@@ -6,7 +6,7 @@
 
 from azure.cli.core.commands.parameters import tags_type, get_three_state_flag
 from azure.cli.command_modules.monitor.actions import get_period_type
-from azure.cli.command_modules.monitor.validators import get_action_group_validator
+from azext_scheduled_query._validators import get_action_group_validator
 from knack.arguments import CLIArgumentType
 from ._actions import ScheduleQueryConditionAction, ScheduleQueryAddAction, ScheduleQueryConditionQueryAction
 
@@ -35,4 +35,14 @@ def load_arguments(self, _):
         c.argument('mute_actions_duration', type=get_period_type(as_timedelta=True),
                    options_list=['--mute-actions-duration', '--mad'],
                    help='Mute actions for the chosen period of time (in ISO 8601 duration format) after the alert is fired.')
-        c.argument('actions', options_list=['--action', '-a'], action=ScheduleQueryAddAction, nargs='+', validator=get_action_group_validator('actions'))
+        c.argument('actions', options_list=['--action', '-a'], action=ScheduleQueryAddAction, nargs='+',
+                   validator=get_action_group_validator('actions'), deprecate_info=c.deprecate(redirect='--actions', hide=True))
+        c.argument('action_groups', options_list=['--action-groups'], nargs='+', help='Action Group resource Ids to invoke when the alert fires.')
+        c.argument('custom_properties', options_list=['--custom-properties'], nargs='*', help='The properties of an alert payload.')
+        c.argument('auto_mitigate', arg_type=get_three_state_flag(),
+                   help='The flag that indicates whether the alert should be automatically resolved or not. The default is true.')
+        c.argument('skip_query_validation', arg_type=get_three_state_flag(),
+                   help='The flag which indicates whether the provided query should be validated or not.')
+        c.argument('check_workspace_alerts_storage', options_list=['--check-ws-alerts-storage', '--cwas'],
+                   arg_type=get_three_state_flag(),
+                   help="The flag which indicates whether this scheduled query rule should be stored in the customer's storage.")