+++ src/scheduled-query/azext_scheduled_query/vendored_sdks/azure_mgmt_scheduled_query/models/_models.py
@@ -420,7 +420,7 @@ class ScheduledQueryRuleResource(TrackedResource):
      format) after the alert is fired. Relevant only for rules of the kind LogAlert.
     :type mute_actions_duration: ~datetime.timedelta
     :param actions: Actions to invoke when the alert fires.
-    :type actions: any
+    :type actions: ~$(python-base-namespace).v2021_02_preview.models.Actions
     :ivar is_workspace_alerts_storage_configured: The flag which indicates whether this scheduled
      query rule has been configured to be stored in the customer's storage. The default is false.
     :vartype is_workspace_alerts_storage_configured: bool
@@ -461,7 +461,7 @@ class ScheduledQueryRuleResource(TrackedResource):
         'is_legacy_log_analytics_rule': {'key': 'properties.isLegacyLogAnalyticsRule', 'type': 'bool'},
         'description': {'key': 'properties.description', 'type': 'str'},
         'display_name': {'key': 'properties.displayName', 'type': 'str'},
-        'severity': {'key': 'properties.severity', 'type': 'float'},
+        'severity': {'key': 'properties.severity', 'type': 'int'},
         'enabled': {'key': 'properties.enabled', 'type': 'bool'},
         'scopes': {'key': 'properties.scopes', 'type': '[str]'},
         'evaluation_frequency': {'key': 'properties.evaluationFrequency', 'type': 'duration'},
@@ -470,7 +470,7 @@ class ScheduledQueryRuleResource(TrackedResource):
         'target_resource_types': {'key': 'properties.targetResourceTypes', 'type': '[str]'},
         'criteria': {'key': 'properties.criteria', 'type': 'ScheduledQueryRuleCriteria'},
         'mute_actions_duration': {'key': 'properties.muteActionsDuration', 'type': 'duration'},
-        'actions': {'key': 'properties.actions', 'type': 'object'},
+        'actions': {'key': 'properties.actions', 'type': 'Actions'},
         'is_workspace_alerts_storage_configured': {'key': 'properties.isWorkspaceAlertsStorageConfigured', 'type': 'bool'},
         'check_workspace_alerts_storage_configured': {'key': 'properties.checkWorkspaceAlertsStorageConfigured', 'type': 'bool'},
         'skip_query_validation': {'key': 'properties.skipQueryValidation', 'type': 'bool'},
@@ -569,7 +569,7 @@ class ScheduledQueryRuleResourcePatch(msrest.serialization.Model):
      format) after the alert is fired. Relevant only for rules of the kind LogAlert.
     :type mute_actions_duration: ~datetime.timedelta
     :param actions: Actions to invoke when the alert fires.
-    :type actions: any
+    :type actions: ~$(python-base-namespace).v2021_02_preview.models.Actions
     :ivar is_workspace_alerts_storage_configured: The flag which indicates whether this scheduled
      query rule has been configured to be stored in the customer's storage. The default is false.
     :vartype is_workspace_alerts_storage_configured: bool
@@ -597,7 +597,7 @@ class ScheduledQueryRuleResourcePatch(msrest.serialization.Model):
         'is_legacy_log_analytics_rule': {'key': 'properties.isLegacyLogAnalyticsRule', 'type': 'bool'},
         'description': {'key': 'properties.description', 'type': 'str'},
         'display_name': {'key': 'properties.displayName', 'type': 'str'},
-        'severity': {'key': 'properties.severity', 'type': 'float'},
+        'severity': {'key': 'properties.severity', 'type': 'int'},
         'enabled': {'key': 'properties.enabled', 'type': 'bool'},
         'scopes': {'key': 'properties.scopes', 'type': '[str]'},
         'evaluation_frequency': {'key': 'properties.evaluationFrequency', 'type': 'duration'},
@@ -606,7 +606,7 @@ class ScheduledQueryRuleResourcePatch(msrest.serialization.Model):
         'target_resource_types': {'key': 'properties.targetResourceTypes', 'type': '[str]'},
         'criteria': {'key': 'properties.criteria', 'type': 'ScheduledQueryRuleCriteria'},
         'mute_actions_duration': {'key': 'properties.muteActionsDuration', 'type': 'duration'},
-        'actions': {'key': 'properties.actions', 'type': 'object'},
+        'actions': {'key': 'properties.actions', 'type': 'Actions'},
         'is_workspace_alerts_storage_configured': {'key': 'properties.isWorkspaceAlertsStorageConfigured', 'type': 'bool'},
         'check_workspace_alerts_storage_configured': {'key': 'properties.checkWorkspaceAlertsStorageConfigured', 'type': 'bool'},
         'skip_query_validation': {'key': 'properties.skipQueryValidation', 'type': 'bool'},